Fixed some instructions

+++ src/core/cpu/arm/functions/mul.rs
@@ -19,7 +19,7 @@ pub fn arm_fn_smull(lhs: u32, rhs: u32, _: u32, _: u32) -> u64 {
 }
 
 pub fn arm_fn_smlal(lhs: u32, rhs: u32, acc_hi: u32, acc_lo: u32) -> u64 {
-	((lhs as i64) * (rhs as i64) + (combine32to64(acc_hi, acc_lo) as i64)) as u64
+	(((lhs as i64) * (rhs as i64)) as u64) + combine32to64(acc_hi, acc_lo)
 }
 
 fn combine32to64(hi: u32, lo: u32) -> u64 {
